Update user.ts to be user.util.ts to match naming convention. Add the updateUserAddress api call function to the user.util.ts.
